refactor: strategy pattern for deckfile fetching (#84)
* refactor: strategy pattern for deckfile fetching
- preparation for bugfixes + features regarding deckfile, e.g. #81
- decouple deckfile fetching from parsing
- tests added
* fix: test gefyra demo
* fix: add ssh agent for testing
* fix: increase try it yourself example timeout

+++ getdeck/api/hosts.py
@@ -1,4 +1,5 @@
 import logging
+import os
 import shutil
 import socket
 
@@ -53,8 +54,11 @@ def run_hosts(
             logger.error(f"Unknown host action '{host_action}'")
     else:
         logger.info("No hosts specified in Deckfile")
-    if is_temp_dir:
+
+    # TODO: refactor/remove?
+    if is_temp_dir and os.path.isdir(working_dir_path):
         shutil.rmtree(working_dir_path)
+
     return True
