+++ getdeck/api/remove.py
@@ -1,4 +1,5 @@
 import logging
+import os
 import shutil
 from typing import Callable
 
@@ -25,8 +26,11 @@ def remove_cluster(
         k8s_provider.delete()
     else:
         logger.info("Cluster does not exist")
-    if is_temp_dir:
+
+    # TODO: refactor/remove?
+    if is_temp_dir and os.path.isdir(working_dir_path):
         shutil.rmtree(working_dir_path)
+
     return True
 
 
@@ -83,6 +87,8 @@ def remove_deck(
     else:
         logger.info("Cluster does not exist")
 
-    if is_temp_dir:
+    # TODO: refactor/remove?
+    if is_temp_dir and os.path.isdir(working_dir_path):
         shutil.rmtree(working_dir_path)
+
     return True
